refactor(ismp): use u64 height
Destination state machine could use larger type, so standardise on the
same type provided by ismp.

+++ pallets/api-vnext/src/messaging/precompiles/interfaces/v0/IISMP.sol
@@ -90,7 +90,7 @@ interface IISMP {
         /// @custom:property The destination state machine of this request.
         uint32 destination;
         /// @custom:property Height at which to read the state machine.
-        uint32 height;
+        uint64 height;
         /// @custom:property Relative from the current timestamp at which this request expires in seconds.
         uint64 timeout;
         /// @custom:property Some application-specific metadata relating to this request.
